moving localstorage to services
